+++ sphinx/ext/autodoc/__init__.py
@@ -32,12 +32,7 @@ from sphinx.util.inspect import (
     safe_getattr,
     stringify_signature,
 )
-from sphinx.util.typing import (
-    RenderMode,
-    get_type_hints,
-    restify,
-    stringify_annotation,
-)
+from sphinx.util.typing import get_type_hints, restify, stringify_annotation
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator, Sequence
@@ -49,7 +44,7 @@ if TYPE_CHECKING:
     from sphinx.environment import BuildEnvironment, _CurrentDocument
     from sphinx.events import EventManager
     from sphinx.ext.autodoc.directive import DocumenterBridge
-    from sphinx.util.typing import ExtensionMetadata, OptionSpec, _RestifyRenderMode
+    from sphinx.util.typing import ExtensionMetadata, OptionSpec, _RestifyMode
 
     _AutodocObjType = Literal[
         'module', 'class', 'exception', 'function', 'method', 'attribute'
@@ -82,18 +77,11 @@ special_member_re = re.compile(r'^__\S+__$')
 
 
 def _get_render_mode(
-    config: Config,
-    default: _RestifyRenderMode = RenderMode.fully_qualified_except_typing,
-) -> _RestifyRenderMode:
-    if config.autodoc_typehints_format == 'short':
-        mode: _RestifyRenderMode = RenderMode.smart
-    else:
-        mode = default
-
-    if config.python_display_short_literal_types:
-        mode |= RenderMode.short_literal
-
-    return mode
+    typehints_format: Literal['fully-qualified', 'short'],
+) -> _RestifyMode:
+    if typehints_format == 'short':
+        return 'smart'
+    return 'fully-qualified-except-typing'
 
 
 def identity(x: Any) -> Any:
@@ -576,8 +564,8 @@ class Documenter:
             retann = self.retann
         else:
             # try to introspect the signature
-            retann = None
             try:
+                retann = None
                 args = self._call_format_args(**kwargs)
                 if args:
                     matched = re.match(r'^(\(.*\))\s+->\s+(.*)$', args)
@@ -1493,7 +1481,7 @@ class FunctionDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # typ
         if self.config.autodoc_typehints_format == 'short':
             kwargs.setdefault('unqualified_typehints', True)
         if self.config.python_display_short_literal_types:
-            kwargs.setdefault('short_literal_types', True)
+            kwargs.setdefault('short_literals', True)
 
         try:
             self._events.emit('autodoc-before-process-signature', self.object, False)
@@ -1530,7 +1518,7 @@ class FunctionDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # typ
         if self.config.autodoc_typehints_format == 'short':
             kwargs.setdefault('unqualified_typehints', True)
         if self.config.python_display_short_literal_types:
-            kwargs.setdefault('short_literal_types', True)
+            kwargs.setdefault('short_literals', True)
 
         sigs = []
         if (
@@ -1819,7 +1807,7 @@ class ClassDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # type: 
         if self.config.autodoc_typehints_format == 'short':
             kwargs.setdefault('unqualified_typehints', True)
         if self.config.python_display_short_literal_types:
-            kwargs.setdefault('short_literal_types', True)
+            kwargs.setdefault('short_literals', True)
 
         try:
             self._signature_class, _signature_method_name, sig = self._get_signature()
@@ -1862,7 +1850,7 @@ class ClassDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # type: 
         if self.config.autodoc_typehints_format == 'short':
             kwargs.setdefault('unqualified_typehints', True)
         if self.config.python_display_short_literal_types:
-            kwargs.setdefault('short_literal_types', True)
+            kwargs.setdefault('short_literals', True)
 
         sig = super().format_signature()
         sigs = []
@@ -1959,8 +1947,8 @@ class ClassDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # type: 
                 'autodoc-process-bases', self.fullname, self.object, self.options, bases
             )
 
-            mode = _get_render_mode(self.config)
-            base_classes = [restify(cls, mode) for cls in bases]
+            mode = _get_render_mode(self.config.autodoc_typehints_format)
+            base_classes = [restify(cls, mode=mode) for cls in bases]
 
             sourcename = self.get_sourcename()
             self.add_line('', sourcename)
@@ -2073,19 +2061,21 @@ class ClassDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # type: 
             return None
 
     def add_content(self, more_content: StringList | None) -> None:
-        mode = _get_render_mode(self.config)
+        mode = _get_render_mode(self.config.autodoc_typehints_format)
+        short_literals = self.config.python_display_short_literal_types
 
         if isinstance(self.object, NewType):
-            supertype = restify(self.object.__supertype__, mode)
+            supertype = restify(self.object.__supertype__, mode=mode)
 
             more_content = StringList([_('alias of %s') % supertype, ''], source='')
         if isinstance(self.object, TypeVar):
-            attrs = [
-                repr(self.object.__name__),
-                *(stringify_annotation(c, mode) for c in self.object.__constraints__),
-            ]
+            attrs = [repr(self.object.__name__)]
+            attrs.extend(
+                stringify_annotation(constraint, mode, short_literals=short_literals)
+                for constraint in self.object.__constraints__
+            )
             if self.object.__bound__:
-                bound = restify(self.object.__bound__, mode)
+                bound = restify(self.object.__bound__, mode=mode)
                 attrs.append(r'bound=\ ' + bound)
             if self.object.__covariant__:
                 attrs.append('covariant=True')
@@ -2105,7 +2095,7 @@ class ClassDocumenter(DocstringSignatureMixin, ModuleLevelDocumenter):  # type: 
 
         if self.doc_as_attr and not self.get_variable_comment():
             try:
-                alias = restify(self.object, mode)
+                alias = restify(self.object, mode=mode)
                 more_content = StringList([_('alias of %s') % alias], source='')
             except AttributeError:
                 pass  # Invalid class object is passed.
@@ -2197,8 +2187,8 @@ class GenericAliasMixin(DataDocumenterMixinBase):
 
     def update_content(self, more_content: StringList) -> None:
         if inspect.isgenericalias(self.object):
-            mode = _get_render_mode(self.config)
-            alias = restify(self.object, mode)
+            mode = _get_render_mode(self.config.autodoc_typehints_format)
+            alias = restify(self.object, mode=mode)
 
             more_content.append(_('alias of %s') % alias, '')
             more_content.append('', '')
@@ -2322,19 +2312,13 @@ class DataDocumenter(
                     include_extras=True,
                 )
                 if self.objpath[-1] in annotations:
-                    mode = _get_render_mode(self.config)
+                    mode = _get_render_mode(self.config.autodoc_typehints_format)
+                    short_literals = self.config.python_display_short_literal_types
                     objrepr = stringify_annotation(
-                        annotations.get(self.objpath[-1]), mode
+                        annotations.get(self.objpath[-1]),
+                        mode,
+                        short_literals=short_literals,
                     )
-                    if self.config.autodoc_typehints_format == 'short':
-                        objrepr = stringify_annotation(
-                            annotations.get(self.objpath[-1]), 'smart'
-                        )
-                    else:
-                        objrepr = stringify_annotation(
-                            annotations.get(self.objpath[-1]),
-                            'fully-qualified-except-typing',
-                        )
                     self.add_line('   :type: ' + objrepr, sourcename)
 
             try:
@@ -2425,7 +2409,7 @@ class MethodDocumenter(DocstringSignatureMixin, ClassLevelDocumenter):  # type: 
         if self.config.autodoc_typehints_format == 'short':
             kwargs.setdefault('unqualified_typehints', True)
         if self.config.python_display_short_literal_types:
-            kwargs.setdefault('short_literal_types', True)
+            kwargs.setdefault('short_literals', True)
 
         try:
             if self.object == object.__init__ and self.parent != object:  # NoQA: E721
@@ -2496,7 +2480,7 @@ class MethodDocumenter(DocstringSignatureMixin, ClassLevelDocumenter):  # type: 
         if self.config.autodoc_typehints_format == 'short':
             kwargs.setdefault('unqualified_typehints', True)
         if self.config.python_display_short_literal_types:
-            kwargs.setdefault('short_literal_types', True)
+            kwargs.setdefault('short_literals', True)
 
         sigs = []
         if (
@@ -2980,19 +2964,13 @@ class AttributeDocumenter(  # type: ignore[misc]
                     include_extras=True,
                 )
                 if self.objpath[-1] in annotations:
-                    mode = _get_render_mode(self.config)
+                    mode = _get_render_mode(self.config.autodoc_typehints_format)
+                    short_literals = self.config.python_display_short_literal_types
                     objrepr = stringify_annotation(
-                        annotations.get(self.objpath[-1]), mode
+                        annotations.get(self.objpath[-1]),
+                        mode,
+                        short_literals=short_literals,
                     )
-                    if self.config.autodoc_typehints_format == 'short':
-                        objrepr = stringify_annotation(
-                            annotations.get(self.objpath[-1]), 'smart'
-                        )
-                    else:
-                        objrepr = stringify_annotation(
-                            annotations.get(self.objpath[-1]),
-                            'fully-qualified-except-typing',
-                        )
                     self.add_line('   :type: ' + objrepr, sourcename)
 
             try:
@@ -3031,7 +3009,6 @@ class AttributeDocumenter(  # type: ignore[misc]
         if comment:
             return [comment]
 
-        orig = self.config.autodoc_inherit_docstrings
         try:
             # Disable `autodoc_inherit_docstring` temporarily to avoid to obtain
             # a docstring from the value which descriptor returns unexpectedly.
@@ -3128,14 +3105,11 @@ class PropertyDocumenter(DocstringStripSignatureMixin, ClassLevelDocumenter):  #
                 func, type_aliases=self.config.autodoc_type_aliases
             )
             if signature.return_annotation is not Parameter.empty:
-                mode = _get_render_mode(self.config)
-                objrepr = stringify_annotation(signature.return_annotation, mode)
-                if self.config.autodoc_typehints_format == 'short':
-                    objrepr = stringify_annotation(signature.return_annotation, 'smart')
-                else:
-                    objrepr = stringify_annotation(
-                        signature.return_annotation, 'fully-qualified-except-typing'
-                    )
+                mode = _get_render_mode(self.config.autodoc_typehints_format)
+                short_literals = self.config.python_display_short_literal_types
+                objrepr = stringify_annotation(
+                    signature.return_annotation, mode, short_literals=short_literals
+                )
                 self.add_line('   :type: ' + objrepr, sourcename)
         except TypeError as exc:
             logger.warning(