+++ sphinx/ext/napoleon/docstring.py
@@ -12,14 +12,13 @@ from typing import TYPE_CHECKING, Any
 
 from sphinx.locale import _, __
 from sphinx.util import logging
-from sphinx.util.typing import RenderMode, get_type_hints, stringify_annotation
+from sphinx.util.typing import get_type_hints, stringify_annotation
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator
 
     from sphinx.application import Sphinx
     from sphinx.config import Config as SphinxConfig
-    from sphinx.util.typing import _StringifyMode
 
 logger = logging.getLogger(__name__)
 
@@ -1096,12 +1095,14 @@ class GoogleDocstring:
                     )
                     self._annotations = get_type_hints(self._obj, None, localns)
                 if _name in self._annotations:
-                    mode: _StringifyMode = RenderMode.fully_qualified_except_typing
-                    if getattr(
-                        self._config, 'python_display_short_literal_types', None
-                    ):
-                        mode = RenderMode.fully_qualified_except_typing_short_literal
-                    return stringify_annotation(self._annotations[_name], mode)
+                    short_literals = getattr(
+                        self._config, 'python_display_short_literal_types', False
+                    )
+                    return stringify_annotation(
+                        self._annotations[_name],
+                        mode='fully-qualified-except-typing',
+                        short_literals=short_literals,
+                    )
         # No annotation found
         return ''